新增功能-使用道具、查詢角色基本屬性

+++ backend/database/tables/UserInfo.js
@@ -8,8 +8,8 @@ const createUserInfoTable = async () => {
       LeftOfPlay INT DEFAULT 5,
       TimeCoin DOUBLE DEFAULT 0,
       BaseLeftOfPlay INT DEFAULT 5, -- 基礎可玩次數
-      BaseAttackPower DOUBLE DEFAULT 1, -- 玩家基礎攻擊力
-      RewardMultiplier DOUBLE DEFAULT 1, -- 玩家結算獎勵倍數
+      BaseAttackPower DECIMAL(10, 5) DEFAULT 1, -- 玩家基礎攻擊力
+      RewardMultiplier DECIMAL(10, 5) DEFAULT 1, -- 玩家結算獎勵倍數
       CreatedAt TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
       Creator VARCHAR(255) NOT NULL
     )
